% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asJsonldDataID.R
\name{asJsonldDataID}
\alias{asJsonldDataID}
\title{asJsonldDataID}
\usage{
asJsonldDataID(
  x,
  context = "https://downloads.dbpedia.org/databus/context.jsonld",
  dBusUri = "https://energy.databus.dbpedia.org"
)
}
\arguments{
\item{x}{a named list that contains the metadata and link to the dataset
that will be published on the Databus}

\item{context}{string containing the "context" URI of the Databus}

\item{dBusUri}{string containing the base Databus URI for POST requests}
}
\value{
the DataID documentation as JSON-LD string
}
\description{
Returns the DataID documentation as JSON-LD string
}
\seealso{
\code{\link{asJsonldGroup}}
}
\author{
Anastasis Giannousakis
}
