% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asJsonldGroup.R
\name{asJsonldGroup}
\alias{asJsonldGroup}
\title{asJsonldGroup}
\usage{
asJsonldGroup(
  x,
  context = "https://downloads.dbpedia.org/databus/context.jsonld"
)
}
\arguments{
\item{x}{a named list that contains the metadata of the Group that will
publish data on the Databus}

\item{context}{string containing the "context" URI of the Databus}
}
\value{
the Group documentation as JSON-LD string
}
\description{
Returns the Group documentation as JSON-LD string
}
\seealso{
\code{\link{asJsonldDataID}}
}
\author{
Anastasis Giannousakis
}
