% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databusUpload.R
\name{databusUpload}
\alias{databusUpload}
\title{databusUpload}
\usage{
databusUpload(myKey, myData = NULL)
}
\arguments{
\item{myKey}{a personal and secret API key (can be found or generated
under https://energy.databus.dbpedia.org/{{user}}#settings)}

\item{myData}{a list or a path to a JSON file with the dataset's metadata
(if missing, defaults from the package's inst folder will be taken)}
}
\description{
Publishes a dataset on the Databus (energy.databus.dbpedia.org)
}
\author{
Anastasis Giannousakis
}
