% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfiles_to_dict.R
\name{dbfiles_to_dict}
\alias{dbfiles_to_dict}
\title{dbfiles_to_dict}
\usage{
dbfiles_to_dict(x, artifact, dataid_uri)
}
\arguments{
\item{x}{an unnamed list of lists with mandatory names in each of its items: "metadata" and "file"}

\item{artifact}{string containing the artifact name}

\item{dataid_uri}{string containing the URI used in the POST request that publishes the DataID documentation, as created by \code{\link{getTargetURIDataID}}}
}
\value{
a named list with special strings for file annotation on the Databus
}
\description{
Returns a named list with special strings for file annotation on the Databus
}
\seealso{
\code{\link{getDatabusFile}}
}
\author{
Anastasis Giannousakis
}
