% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_to_dev_databus.R
\name{deploy_to_dev_databus}
\alias{deploy_to_dev_databus}
\title{deploy_to_dev_databus}
\usage{
deploy_to_dev_databus(
  api_key,
  group,
  dataid,
  post_databus_uri = "https://energy.databus.dbpedia.org/system/publish"
)
}
\arguments{
\item{api_key}{a personal and secret API key (can be found or generated
under https://energy.databus.dbpedia.org/{{user}}#settings)}

\item{group}{the Group documentation as JSON-LD string}

\item{dataid}{the DataID documentation as JSON-LD string}

\item{post_databus_uri}{string containing the Databus URI for POST requests}
}
\value{
the POST request response object for further analysis
}
\description{
Function that executes the deployment to the Databus (energy.databus.dbpedia.org)
}
\author{
Anastasis Giannousakis
}
