% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_to_energy_databus.R
\name{deploy_to_energy_databus}
\alias{deploy_to_energy_databus}
\title{deploy_to_energy_databus}
\usage{
deploy_to_energy_databus(api_key, groupJson = NULL, dataidJson = NULL)
}
\arguments{
\item{api_key}{a personal and secret API key (can be found or generated
under https://energy.databus.dbpedia.org/{{user}}#settings)}

\item{groupJson}{the Group documentation as JSON-LD string (if missing defaults
from the package inst folder will be taken)}

\item{dataidJson}{the DataID documentation as JSON-LD string (if missing defaults
from the package inst folder will be taken)}
}
\description{
Function that executes the deployment to the Databus (energy.databus.dbpedia.org)
}
\author{
Anastasis Giannousakis
}
