% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatabusFile.R
\name{getDatabusFile}
\alias{getDatabusFile}
\title{getDatabusFile}
\usage{
getDatabusFile(x)
}
\arguments{
\item{x}{a list that has to contain the named items "file", "metadata" and "filetype"}
}
\value{
a named list with the information gathered
}
\description{
Fetches the necessary information of a file URI for the deployment to the Databus
}
\author{
Anastasis Giannousakis
}
