% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTargetURIDataID.R
\name{getTargetURIDataID}
\alias{getTargetURIDataID}
\title{getTargetURIDataID}
\usage{
getTargetURIDataID(x, dBusUri = "https://energy.databus.dbpedia.org")
}
\arguments{
\item{x}{a list that has to contain the named strings "account_name", "group", "artifact" and "version"}

\item{dBusUri}{string containing the base Databus URI for POST requests}
}
\value{
the URI used in the POST request that publishes the DataID documentation
}
\description{
Returns the URI used in the POST request that publishes the DataID documentation
}
\seealso{
\code{\link{deploy_to_dev_databus}}
}
\author{
Anastasis Giannousakis
}
