% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTargetURIGroup.R
\name{getTargetURIGroup}
\alias{getTargetURIGroup}
\title{getTargetURIGroup}
\usage{
getTargetURIGroup(x, dBusUri = "https://dev.databus.dbpedia.org")
}
\arguments{
\item{x}{a list that has to contain the named strings "account_name" and "id"}

\item{dBusUri}{string containing the base Databus URI for POST requests}
}
\value{
the URI used in the POST request that publishes Group documentation
}
\description{
Returns the URI used in the POST request that publishes the Group documentation
}
\seealso{
\code{\link{deploy_to_dev_databus}}
}
\author{
Anastasis Giannousakis
}
