% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2databus-package.R
\docType{package}
\name{r2databus-package}
\alias{r2databus}
\alias{r2databus-package}
\title{r2databus: An R interface for the DBpedia Databus}
\description{
Provides tools for uploading, downloading, and annotating data to the DBpedia Databus.
}
\author{
\strong{Maintainer}: Anastasis Giannousakis \email{giannou@pik-potsdam.de}

}
