% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databusQuery.R
\name{databusQuery}
\alias{databusQuery}
\title{databusQuery}
\usage{
databusQuery(
  url = "https://databus.dbpedia.org/repo/sparql",
  q = NULL,
  metadata = list(`dataid-cv:type` = c("emissiondata", "bla"))
)
}
\arguments{
\item{url}{string containing the Databus URL for SPARQL queries (endpoint)}

\item{q}{the SPARQL query (prefixes and query code)}

\item{metadata}{the metadata to search for (list of key-value(s) pairs, see example)}
}
\value{
the SPARQL query response object for further analysis
}
\description{
Function that executes a SPARQL Query on the Databus
}
\author{
Anastasis Giannousakis
}
