% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDatabusDownload.R
\name{toolDatabusDownload}
\alias{toolDatabusDownload}
\title{Tool: DatabusDownload}
\usage{
toolDatabusDownload(url, subtype = NULL)
}
\arguments{
\item{url}{the Databus page hosting the metadata}

\item{subtype}{the specific dataset to download}
}
\value{
a list with the metadata associated with the dataset
}
\description{
This function downloads a dataset and its metadata from the Energy Databus (https://energy.databus.dbpedia.org/)
}
\examples{
toolDatabusDownload("https://energy.databus.dbpedia.org/ludee/dea-technology-data/rli-dea-td-generation-wind-turbine",
subtype = "https/2022-10-12/rli-dea-td-generation-wind-turbine.zip")

}
\author{
Anastasis Giannousakis
}
